requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
openai==1.61.1
langchain_openai=0.3.5
langchain==0.3.18
itsdangerous==2.2.0
python-dotenv== 1.0.1